+++ packages/iov-crypto/types/index.d.ts
@@ -2,7 +2,6 @@ export * from "./bip39";
 export * from "./hmac";
 export * from "./keccak";
 export * from "./libsodium";
-export * from "./ripemd160";
 export * from "./secp256k1";
 export { ExtendedSecp256k1Signature, Secp256k1Signature } from "./secp256k1signature";
 export * from "./sha";